requirements.txt
fastapi
uvicorn
shapely
pyproj
requests
langchain
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
openai
python-dotenv
numpy
